Added tim/img files to build the secondary image
A3700 bootROM supports booting from the SPI flash at the
address 0x200000 if the data on header of the SPI flash
is corrupted. Creating the secondary flash-image will need
to setup the tim.txt and img-*.txt with the correct offset.

This commit is to checkin the tim/img-1/img-2 files with
the 2MB offset which are needed to build the secondary
boot flash image.

Also included a readme file with descriptions on how to
build and burn the secondary boot image.

Change-Id: Ib6364ece2b257a4c4b1b864341ab04b0984226bc